% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_city.R
\name{add_city}
\alias{add_city}
\title{Main escape city creation function: take in city name, scrape tables, write tables to database}
\usage{
add_city(city)
}
\arguments{
\item{city}{String of the city name wished to add to escape database}
}
\description{
Main escape city creation function: take in city name, scrape tables, write tables to database
}

