% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCell.R
\name{check_genes}
\alias{check_genes}
\title{Check if all genes in signatures are found in data matrix - otherwise add zero counts in data-matrix to complete it}
\usage{
check_genes(matrix, features)
}
\arguments{
\item{matrix}{Input data matrix}

\item{features}{List of genes that must be present (otherwise they are added)}
}
\value{
Same input matrix, extended to comprise any missing genes
}
\description{
Check if all genes in signatures are found in data matrix - otherwise add zero counts in data-matrix to complete it
}
