% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCell.R
\name{check_signature_names}
\alias{check_signature_names}
\title{Check signature names and add standard names is missing}
\usage{
check_signature_names(features)
}
\arguments{
\item{features}{List of signatures for scoring}
}
\description{
Check signature names and add standard names is missing
}
