% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_DM.R
\name{create_dist_matrix}
\alias{create_dist_matrix}
\title{Function to create distance matrix between pairs of every escape location in city}
\usage{
create_dist_matrix(activity, food)
}
\arguments{
\item{activity}{Data frame returned from activity scraping function}

\item{food}{Data frame returned from food scraping function}
}
\description{
Function to create distance matrix between pairs of every escape location in city
}

