% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCell.R
\name{data_to_ranks_data_table}
\alias{data_to_ranks_data_table}
\title{Calculate per-cell feature rankings}
\usage{
data_to_ranks_data_table(data, ties.method = "average")
}
\arguments{
\item{data}{Expression data matrix}

\item{ties.method}{How to break ties (passed on to data.table::frankv)}
}
\description{
Calculate per-cell feature rankings
}
