% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_DM.R
\name{dist}
\alias{dist}
\title{Function to calculate distance between two locations, given their unique event ids}
\usage{
dist(id1, id2)
}
\arguments{
\item{id1}{event id for first event in pair}

\item{id2}{event id for second event in pair}
}
\description{
Function to calculate distance between two locations, given their unique event ids
}

