% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape.gene.sets.R
\docType{data}
\name{escape.gene.sets}
\alias{escape.gene.sets}
\title{Built-In Gene Sets for escape}
\description{
A list of gene sets derived from Azizi, et al 2018 
\href{https://pubmed.ncbi.nlm.nih.gov/29961579/}{PMID: 29961579})
relating to tumor immunity.
}
