% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_activity.R
\name{generate_activities}
\alias{generate_activities}
\title{Function to scrape web for activities in a given city, returns 'activity' dataframe}
\usage{
generate_activities(city)
}
\arguments{
\item{city}{A string representing the city for which you want activities.}
}
\value{
A dataframe with all of the activies for the input city.
}
\description{
Function to scrape web for activities in a given city, returns 'activity' dataframe
}

