% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{masterPCAPlot}
\alias{masterPCAPlot}
\title{Visualize the components of the PCA analysis of the enrichment results}
\usage{
masterPCAPlot(enriched, PCx, PCy, top.contribution = 10)
}
\arguments{
\item{enriched}{The output of \code{\link{enrichIt}}.}

\item{PCx}{The principal component graphed on the x-axis.}

\item{PCy}{The principal component graphed on the y-axis.}

\item{top.contribution}{The number of gene sets to graph, organized 
by PCA contribution.}
}
\value{
ggplot2 object sumamrizing the PCA for the enrichment scores
}
\description{
Graph the major gene set contributors to the \code{\link{pcaEnrichment}}.
}
\examples{
ES2 <- readRDS(url(
"https://ncborcherding.github.io/vignettes/escape_enrichment_results.rds"))
masterPCAPlot(ES2, PCx = "PC1", PCy = "PC2", top.contribution = 10)

}
\seealso{
\code{\link{enrichIt}} for generating enrichment scores.
}
