% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCell.R
\name{rankings2Uscore}
\alias{rankings2Uscore}
\title{Get signature scores from pre-computed rank matrix}
\usage{
rankings2Uscore(
  ranks_matrix,
  features,
  chunk.size = 1000,
  w_neg = 1,
  ncores = 1,
  force.gc = FALSE,
  name = "_UCell"
)
}
\arguments{
\item{ranks_matrix}{A rank matrix}

\item{features}{List of signatures}

\item{chunk.size}{How many cells per matrix chunk}

\item{w_neg}{Weight on negative signatures}

\item{ncores}{How many cores to use for parallelization}

\item{force.gc}{Force garbage collection to recover RAM? (FALSE)}

\item{name}{Name suffix for metadata columns ("_UCell")}
}
\value{
A list of signature scores
}
\description{
Get signature scores from pre-computed rank matrix
}
