% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCell.R
\name{StoreRankings_UCell}
\alias{StoreRankings_UCell}
\title{Calculate and store gene rankings for a single-cell dataset}
\usage{
StoreRankings_UCell(
  matrix,
  maxRank = 1500,
  chunk.size = 1000,
  ncores = 1,
  assay = "counts",
  ties.method = "average",
  force.gc = FALSE,
  seed = 123
)
}
\arguments{
\item{matrix}{Input matrix, either stored in a [SingleCellExperiment] object or as a raw matrix. \code{dgCMatrix} format supported.}

\item{maxRank}{Maximum number of genes to rank per cell; above this rank, a given gene is considered as not expressed}

\item{chunk.size}{Number of cells to be processed simultaneously (lower size requires slightly more computation but reduces memory demands)}

\item{ncores}{Number of processors to parallelize computation}

\item{assay}{Assay where the data is to be found (for input in 'sce' format)}

\item{ties.method}{How ranking ties should be resolved (passed on to [data.table::frank])}

\item{force.gc}{Explicitly call garbage collector to reduce memory footprint}

\item{seed}{Integer seed}
}
\value{
Returns a sparse matrix of pre-calculated ranks that can be used multiple times to evaluate different signatures

Returns a sparse matrix of pre-calculated ranks that can be used multiple times to evaluate different signatures
}
\description{
Calculate and store gene rankings for a single-cell dataset
}
