% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCell.R
\name{u_stat}
\alias{u_stat}
\title{Calculate Mann Whitney U from a vector of ranks}
\usage{
u_stat(rank_value, maxRank = 1000, sparse = FALSE)
}
\arguments{
\item{rank_value}{A vector of ranks}

\item{maxRank}{Max number of features to include in ranking}

\item{sparse}{Whether the vector of ranks is in sparse format}
}
\value{
Normalized AUC (as U statistic) for the vector
}
\description{
Calculate Mann Whitney U from a vector of ranks
}
