% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db_tables.R
\name{write_db_tables}
\alias{write_db_tables}
\title{Function to add scraped dataframes as tables in the escape database}
\usage{
write_db_tables(activity, food, DM, city)
}
\arguments{
\item{activity}{Data frame returned from activity scraping function}

\item{food}{Data frame returned from food scraping function}

\item{DM}{Data frame of distance matrix generated from location data scraped into respective tables}

\item{city}{String of the city name wished to add to escape database}
}
\description{
Function to add scraped dataframes as tables in the escape database
}

