% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityEnrichment.R
\name{densityEnrichment}
\alias{densityEnrichment}
\title{Visualize the mean density ranking of genes across gene set}
\usage{
densityEnrichment(
  input.data,
  gene.set.use,
  gene.sets,
  group.by = NULL,
  palette = "inferno"
)
}
\arguments{
\item{input.data}{A *Seurat* or *SummarizedExperiment* object.}

\item{gene.set.use}{Character(1).  Name of the gene set to display.}

\item{gene.sets}{Named list or `GeneSetCollection` supplying the sets.}

\item{group.by}{Metadata column used to define groups (default `"ident"`).}

\item{palette}{Colour palette from \link[grDevices]{hcl.colors}
(default `"inferno"`).}
}
\value{
A `patchwork`/`ggplot2` object.
}
\description{
This function allows to the user to examine the mean ranking
within the groups across the gene set. The visualization uses
the density function to display the relative position and distribution
of rank.
}
\examples{
GS <- list(Bcells = c("MS4A1", "CD79B", "CD79A", "IGH1", "IGH2"),
           Tcells = c("CD3E", "CD3D", "CD3G", "CD7","CD8A"))
pbmc_small <- SeuratObject::pbmc_small
                        
densityEnrichment(pbmc_small, 
                  gene.set.use = "Tcells",
                  gene.sets = GS)

}
