% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityEnrichment.R
\name{densityEnrichment}
\alias{densityEnrichment}
\title{Visualize the mean density ranking of genes across gene set}
\usage{
densityEnrichment(
  input.data,
  gene.set.use = NULL,
  gene.set.reference = NULL,
  group.by = NULL,
  palette = "inferno"
)
}
\arguments{
\item{input.data}{The single-cell object to use.}

\item{gene.set.use}{Selected individual gene set.}

\item{gene.set.reference}{The gene set library to use to extract 
the individual gene set information from.}

\item{group.by}{Categorical parameter to plot along the x.axis. If input is
a single-cell object the default will be cluster.}

\item{palette}{Colors to use in visualization - input any 
\link[grDevices]{hcl.pals}.}
}
\value{
ggplot2 object mean rank gene density across groups
}
\description{
This function allows to the user to examine the mean ranking
within the groups across the gene set. The visualization uses
the density function to display the relative position and distribution
of rank.
}
\examples{
GS <- list(Bcells = c("MS4A1", "CD79B", "CD79A", "IGH1", "IGH2"),
           Tcells = c("CD3E", "CD3D", "CD3G", "CD7","CD8A"))
pbmc_small <- SeuratObject::pbmc_small
                        
densityEnrichment(pbmc_small, 
                  gene.set.use = "Tcells",
                  gene.set.reference = GS)

}
