% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrichIt}
\alias{enrichIt}
\title{Calculate gene set enrichment scores for single-cell data}
\usage{
enrichIt(
  obj,
  gene.sets = NULL,
  method = "ssGSEA",
  groups = 1000,
  cores = 2,
  min.size = 5
)
}
\arguments{
\item{obj}{The count matrix, Seurat, or SingleCellExperiment object.}

\item{gene.sets}{Gene sets from \code{\link{getGeneSets}} to use 
for the enrichment analysis. Alternatively a simple base R list where
the names of the list elements correspond to the name of the gene set
and the elements themselves are simple vectors of gene names representing
the gene set.}

\item{method}{select the method to calculate enrichment, either "ssGSEA", "UCell" or
"singscore"}

\item{groups}{The number of cells to separate the enrichment calculation.}

\item{cores}{The number of cores to use for parallelization.}

\item{min.size}{Minimum number of gene nessecary to perform the enrichment
calculation}
}
\value{
Data frame of normalized enrichmenet scores (NES)
}
\description{
This function allows users to input both the single-cell RNA-sequencing 
counts and any gene set pathways either from the stored data or from 
other sources. The enrichment calculation itself 
uses the two methods 1) gsva R package and the poisson distribution for RNA
or the \href{https://github.com/carmonalab/UCell}{UCell package}.
}
\examples{
GS <- list(Bcells = c("MS4A1", "CD79B", "CD79A", "IGH1", "IGH2"),
  Tcells = c("CD3E", "CD3D", "CD3G", "CD7","CD8A"))
pbmc_small <- suppressWarnings(SeuratObject::pbmc_small)
ES <- enrichIt(obj = pbmc_small, gene.sets = GS, min.size = NULL)

}
\seealso{
\code{\link{getGeneSets}} to collect gene sets.
}
\author{
Nick Borcherding, Jared Andrews
}
