% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentPlot.R
\name{enrichmentPlot}
\alias{enrichmentPlot}
\title{Gene Rank Enrichment Plot
Display the rank order density for individual gene sets 
by group identify. This function will use the group variable 
to take the mean rank order across all individual cells.}
\usage{
enrichmentPlot(
  obj,
  gene.set,
  gene.sets,
  group,
  colors = c("#0D0887FF", "#7E03A8FF", "#CC4678FF", "#F89441FF", "#F0F921FF")
)
}
\arguments{
\item{obj}{The Seurat or SingleCellExperiment object.}

\item{gene.set}{The name of the specific gene set to visualize}

\item{gene.sets}{Gene sets from \code{\link{getGeneSets}} to use}

\item{group}{The header in the meta data that will be used for the comparison}

\item{colors}{The color palette for the enrichment plot}
}
\value{
ggplot2 object mean rank gene density
}
\description{
Gene Rank Enrichment Plot
Display the rank order density for individual gene sets 
by group identify. This function will use the group variable 
to take the mean rank order across all individual cells.
}
\examples{
 \dontrun{
 GS <- list(Housekeeping = c("ACTA1", "ACTN1", "GAPDH"),
 Cancer = c("TP53","BRCA2","ERBB2","MYC"))
 pbmc_small <- suppressWarnings(SeuratObject::pbmc_small)
 
 enrichmentPlot(pbmc_small gene.set = "Cancer",
                gene.sets = GS, group = "groups")
 }
}
