% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEscape.R
\name{escape.matrix}
\alias{escape.matrix}
\title{Calculate gene set enrichment scores}
\usage{
escape.matrix(
  input.data,
  gene.sets = NULL,
  method = "ssGSEA",
  groups = 1000,
  min.size = 5,
  normalize = FALSE,
  ...
)
}
\arguments{
\item{input.data}{The count matrix, Seurat, or Single-Cell Experiment object.}

\item{gene.sets}{Gene sets can be a list, output from 
\code{\link{getGeneSets}}, or the built-in gene sets 
in the escape package \code{\link{escape.gene.sets}}.}

\item{method}{select the method to calculate enrichment, 
\strong{GSVA}, \strong{ssGSEA} or \strong{UCell}}

\item{groups}{The number of cells to separate the enrichment calculation.}

\item{min.size}{Minimum number of gene necessary to perform the enrichment
calculation}

\item{normalize}{Whether to divide the enrichment score by the number 
of genes \strong{TRUE} or report unnormalized \strong{FALSE}}

\item{...}{pass arguments to GSVA, ssGSEA or UCell call}
}
\value{
Data frame of normalized enrichmenet scores (NES)
}
\description{
This function allows users to input both the single-cell RNA-sequencing 
counts and output the enrichment scores as a matrix.
}
\examples{
GS <- list(Bcells = c("MS4A1", "CD79B", "CD79A", "IGH1", "IGH2"),
           Tcells = c("CD3E", "CD3D", "CD3G", "CD7","CD8A"))
pbmc_small <- SeuratObject::pbmc_small
ES <- escape.matrix(pbmc_small, 
                    gene.sets = GS, 
                    min.size = NULL)

}
\seealso{
\code{\link{getGeneSets}} to collect gene sets.
}
\author{
Nick Borcherding, Jared Andrews
}
