% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneSets.R
\name{getGeneSets}
\alias{getGeneSets}
\title{Get a collection of gene sets from the msigdb}
\usage{
getGeneSets(
  species = "Homo sapiens",
  library = NULL,
  subcategory = NULL,
  gene.sets = NULL,
  version = "7.4",
  id = "SYM"
)
}
\arguments{
\item{species}{The scientific name of the species of interest; only 
"Homo sapiens" or "Mus musculus" are supported.}

\item{library}{A character vector of main collections (e.g. "H", "C5"). 
If provided, only gene sets in these collections are returned.}

\item{subcategory}{A character vector specifying sub-collection abbreviations 
(e.g. "CGP", "CP:REACTOME") to further subset the gene sets.}

\item{gene.sets}{A character vector of specific gene set names to select. 
This filter is applied after other subsetting.}

\item{version}{The version of MSigDB to use (default "7.4").}

\item{id}{The gene identifier type to use (default "SYM" for gene symbols).}
}
\value{
A named list of gene identifiers for each gene set.
}
\description{
This function retrieves gene sets from msigdb and caches the downloaded object 
for future calls. It allows subsetting by main collection (library), 
subcollection, or specific gene sets, and only supports human 
("Homo sapiens") and mouse ("Mus musculus").
}
\examples{
# Get all hallmark gene sets from human.
gs <- getGeneSets(species = "Homo sapiens", 
                  library = "H")

# Get a subset based on main collection and subcollection.
gs <- getGeneSets(species = "Homo sapiens", 
                  library = c("C2", "C5"), 
                  subcategory = "GO:BP")

}
