% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{getGeneSets}
\alias{getGeneSets}
\title{Get a collection of gene sets to perform enrichment on}
\usage{
getGeneSets(
  species = "Homo sapiens",
  library = NULL,
  subcategory = NULL,
  gene.sets = NULL
)
}
\arguments{
\item{species}{The scientific name of the species of interest in 
order to get correcent gene nomenclature}

\item{library}{Individual collection(s) of gene sets, e.g. c("H", "C5").
See \href{https://www.gsea-msigdb.org/gsea/msigdb/collections.jsp}{msigdbr}for
all MSigDB collections.}

\item{subcategory}{MSigDB sub-collection abbreviation, such as CGP or BP.}

\item{gene.sets}{Select gene sets or pathways, using specific names, 
example: pathways = c("HALLMARK_TNFA_SIGNALING_VIA_NFKB"). Will only be
honored if library is set, too.}
}
\value{
A \code{GeneSetCollection} object containing the requested \code{GeneSet} objects.
}
\description{
This function allows users to select libraries and specific 
gene.sets to form a GeneSetCollection that is a list of gene sets.
}
\examples{
GS <- getGeneSets(library = "H")

}
\author{
Nick Borcherding, Jared Andrews
}
