% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{getGeneSets}
\alias{getGeneSets}
\title{Get a collection of gene sets to perform enrichment on}
\usage{
getGeneSets(species = "Homo sapiens", library = NULL, gene.sets = NULL)
}
\arguments{
\item{species}{The scientific name of the species of interest in 
order to get correcent gene nomenclature}

\item{library}{Individual libraries or multiple libraries to select, 
example: library = c("H", "C5").}

\item{gene.sets}{Select gene sets or pathways, using specific names, 
example: pathways = c("HALLMARK_TNFA_SIGNALING_VIA_NFKB").}
}
\value{
List of GeneSets in collection format
}
\description{
This function allows users to select libraries and specific 
gene.sets to form a GeneSetCollection that is a list of gene sets.
}
\examples{
GS <- getGeneSets(library = "H")

}
\author{
Nick Borcherding, Jared Andrews
}
