% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapEnrichment.R
\name{heatmapEnrichment}
\alias{heatmapEnrichment}
\title{Generate a heatmap to visualize enrichment values}
\usage{
heatmapEnrichment(
  input.data,
  assay = NULL,
  group.by = NULL,
  gene.set.use = "all",
  cluster.rows = FALSE,
  cluster.columns = FALSE,
  scale = FALSE,
  facet.by = NULL,
  palette = "inferno"
)
}
\arguments{
\item{input.data}{Enrichment output from \code{\link{escape.matrix}} or
\code{\link{runEscape}}.}

\item{assay}{Name of the assay to plot if data is a single-cell object.}

\item{group.by}{Categorical parameter to plot along the x.axis. If input is
a single-cell object the default will be cluster.}

\item{gene.set.use}{Selected gene sets to visualize. If \strong{"all"}, the 
heatmap will be generated across all gene sets.}

\item{cluster.rows}{Use Euclidean distance to order the row values.}

\item{cluster.columns}{Use Euclidean distance to order the column values.}

\item{scale}{Visualize raw values \strong{FALSE} or Z-transform 
enrichment values \strong{TRUE}.}

\item{facet.by}{Variable to facet the plot into n distinct graphs.}

\item{palette}{Colors to use in visualization - input any 
\link[grDevices]{hcl.pals}.}
}
\value{
ggplot2 object with heatmap of mean enrichment values
}
\description{
This function allows to the user to examine the heatmap with the mean
enrichment values by group. The heatmap will have the gene sets as rows
and columns will be the grouping variable.
}
\examples{
GS <- list(Bcells = c("MS4A1", "CD79B", "CD79A", "IGH1", "IGH2"),
           Tcells = c("CD3E", "CD3D", "CD3G", "CD7","CD8A"))
pbmc_small <- SeuratObject::pbmc_small
pbmc_small <- runEscape(pbmc_small, 
                        gene.sets = GS, 
                        min.size = NULL)
                        
heatmapEnrichment(pbmc_small, 
                  assay = "escape")

}
