% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaEnrichment.R
\name{pcaEnrichment}
\alias{pcaEnrichment}
\title{Visualize the PCA of enrichment values}
\usage{
pcaEnrichment(
  input.data,
  dimRed = NULL,
  x.axis = "PC1",
  y.axis = "PC2",
  facet.by = NULL,
  style = "point",
  add.percent.contribution = TRUE,
  display.factors = FALSE,
  number.of.factors = 10,
  palette = "inferno"
)
}
\arguments{
\item{input.data}{PCA from \code{\link{performPCA}}.}

\item{dimRed}{Name of the dimensional reduction to plot if data is a single-cell object.}

\item{x.axis}{Component to plot on the x.axis.}

\item{y.axis}{Component set to plot on the y.axis.}

\item{facet.by}{Variable to facet the plot into n distinct graphs.}

\item{style}{Return a \strong{"hex"} bin plot or a \strong{"point"}-based plot.}

\item{add.percent.contribution}{Add the relative percent of contribution of the 
selected components to the axis labels.}

\item{display.factors}{Add an arrow overlay to show the direction and magnitude of individual 
gene sets on the PCA dimensions.}

\item{number.of.factors}{The number of gene.sets to display on the overlay.}

\item{palette}{Colors to use in visualization - input any 
\link[grDevices]{hcl.pals}.}
}
\value{
ggplot2 object with PCA distribution
}
\description{
This function allows to the user to examine the distribution
of principal components run on the enrichment values.
}
\examples{
GS <- list(Bcells = c("MS4A1", "CD79B", "CD79A", "IGH1", "IGH2"),
           Tcells = c("CD3E", "CD3D", "CD3G", "CD7","CD8A"))
pbmc_small <- SeuratObject::pbmc_small
pbmc_small <- runEscape(pbmc_small, 
                        gene.sets = GS, 
                        min.size = NULL)
                        
pbmc_small <- performPCA(pbmc_small, 
                         assay = "escape")
                         
pcaEnrichment(pbmc_small,
              x.axis = "PC1",
              y.axis = "PC2",
              dimRed = "escape.PCA")

}
