% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{pcaEnrichment}
\alias{pcaEnrichment}
\title{Density plot of the principal components}
\usage{
pcaEnrichment(
  PCAout,
  PCx,
  PCy,
  colors = c("#0D0887FF", "#7E03A8FF", "#CC4678FF", "#F89441FF", "#F0F921FF"),
  contours = TRUE,
  facet = NULL
)
}
\arguments{
\item{PCAout}{The output of \code{\link{performPCA}}}

\item{PCx}{The principal component graphed on the x-axis}

\item{PCy}{The principal component graphed on the y-axis}

\item{colors}{The color palette for the density plot}

\item{contours}{Binary classifier to add contours to the density plot}

\item{facet}{A parameter to separate the graph}
}
\value{
ggplot2 object of the results of PCA for the enrichment scores
}
\description{
Density plot of the principal components
}
\examples{
ES2 <- readRDS(url(
"https://ncborcherding.github.io/vignettes/escape_enrichment_results.rds"))
PCA <- performPCA(enriched = ES2, groups = c("Type", "Cluster"))
pcaEnrichment(PCA, PCx = "PC1", PCy = "PC2", contours = TRUE)

}
\seealso{
\code{\link{performPCA}} for generating PCA results.
}
