% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performNormalization.R
\name{performNormalization}
\alias{performNormalization}
\title{Perform Normalization on Enrichment Data}
\usage{
performNormalization(
  input.data,
  assay = NULL,
  gene.set.reference = NULL,
  make.positive = TRUE
)
}
\description{
This function allows users to normalize the enrichment
calculations by accounting for single-cell dropout and
producing positive values for downstream differential 
enrichment analyses.
}
