% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performNormalization.R
\name{performNormalization}
\alias{performNormalization}
\title{Perform Normalization on Enrichment Data}
\usage{
performNormalization(
  input.data,
  assay = NULL,
  gene.sets = NULL,
  make.positive = FALSE,
  scale.factor = NULL
)
}
\arguments{
\item{input.data}{Enrichment output from \code{\link{escape.matrix}} or
\code{\link{runEscape}}.}

\item{assay}{Name of the assay to plot if data is a single-cell object.}

\item{gene.sets}{The gene set library to use to extract 
the individual gene set information from.}

\item{make.positive}{Shift enrichment values to a positive range \strong{TRUE}
for downstream analysis or not \strong{TRUE} (default).}

\item{scale.factor}{A vector to use for normalizing enrichment scores per cell.}
}
\value{
Single-cell object or matrix of normalized enrichment scores
}
\description{
This function allows users to normalize the enrichment calculations 
by accounting for single-cell dropout and producing positive 
values for downstream differential enrichment analyses. A positive range
values is useful for several downstream analyses, like differential 
evaluation for log2-fold change, but will alter the original 
enrichment values.
}
\examples{
GS <- list(Bcells = c("MS4A1", "CD79B", "CD79A", "IGH1", "IGH2"),
           Tcells = c("CD3E", "CD3D", "CD3G", "CD7","CD8A"))
pbmc_small <- SeuratObject::pbmc_small
pbmc_small <- runEscape(pbmc_small, 
                        gene.sets = GS, 
                        min.size = NULL)
                        
pbmc_small <- performNormalization(pbmc_small, 
                                   assay = "escape", 
                                   gene.sets = GS)

}
