% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{performPCA}
\alias{performPCA}
\title{Calculate Principal Components for the Enrichment Scores}
\usage{
performPCA(enriched, groups)
}
\arguments{
\item{enriched}{The output of \code{\link{enrichIt}}.}

\item{groups}{The column headers to use in future graphing functions.}
}
\value{
Data frame of principal compoenents
}
\description{
Using all or selected enrichment scores of individual 
single-cells, this function will calculate 
principal components using scaled values and attach 
to the output columns to use to graph later.
}
\examples{
ES2 <- readRDS(url(
"https://ncborcherding.github.io/vignettes/escape_enrichment_results.rds"))
PCA <- performPCA(enriched = ES2, groups = c("Type", "Cluster"))

}
\author{
Nick Borcherding
}
