% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEscape.R
\name{runEscape}
\alias{runEscape}
\title{Ibex single cell calculation}
\usage{
runEscape(
  input.data,
  gene.sets = NULL,
  method = "ssGSEA",
  groups = 1000,
  min.size = 5,
  normalize = FALSE,
  new.assay.name = "escape",
  ...
)
}
\arguments{
\item{gene.sets}{Gene sets can be a list, output from 
\code{\link{getGeneSets}}, or the built-in gene sets 
in the escape package \code{\link{escape.gene.sets}}.}

\item{method}{Select the method to calculate enrichment, \strong{AUCell},
\strong{GSVA}, \strong{ssGSEA} or \strong{UCell}.}

\item{groups}{The number of cells to separate the enrichment calculation.}

\item{min.size}{Minimum number of gene necessary to perform the enrichment
calculation}

\item{normalize}{Whether to divide the enrichment score by the number 
of genes \strong{TRUE} or report unnormalized \strong{FALSE}}

\item{...}{pass arguments to AUCell GSVA, ssGSEA or UCell call}

\item{new.assay.nam}{The new name of the assay to append to 
the single-cell object containing the enrichment scores.}
}
\value{
Seurat or Single-Cell Experiment object with escape enrichment scores
in the assay slot.
}
\description{
Run Ibex algorithm with Seurat or SingleCellExperiment pipelines
Enrichment calculation for single-cell workflows
}
\details{
Run the escape-based gene-set enrichment calculation with 
Seurat or SingleCellExperiment pipelines
}
\examples{
GS <- list(Bcells = c("MS4A1", "CD79B", "CD79A", "IGH1", "IGH2"),
           Tcells = c("CD3E", "CD3D", "CD3G", "CD7","CD8A"))
pbmc_small <- SeuratObject::pbmc_small
pbmc_small <- runEscape(pbmc_small, 
                        gene.sets = GS, 
                        min.size = NULL)

#' @param input.data The count matrix, Seurat, or Single-Cell Experiment object.
}
