% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterEnrichment.R
\name{scatterEnrichment}
\alias{scatterEnrichment}
\title{Generate a density-based scatter plot}
\usage{
scatterEnrichment(
  input.data,
  assay = NULL,
  x.axis = NULL,
  y.axis = NULL,
  scale = FALSE,
  facet.by = NULL,
  style = "point",
  palette = "inferno"
)
}
\arguments{
\item{input.data}{Enrichment output from \code{\link{escape.matrix}} or
\code{\link{runEscape}}.}

\item{assay}{Name of the assay to plot if data is a single-cell object.}

\item{x.axis}{Gene set to plot on the x.axis.}

\item{y.axis}{Gene set to plot on the y.axis.
\strong{group.by} parameter or use the \strong{gene.set} name if wanting to 
apply a gradient palette.}

\item{scale}{Visualize raw values \strong{FALSE} or Z-transform 
enrichment values \strong{TRUE}.}

\item{facet.by}{Variable to facet the plot into n distinct graphs.}

\item{style}{Return a \strong{"hex"} bin plot or a \strong{"point"}-based plot.}

\item{palette}{Colors to use in visualization - input any 
\link[grDevices]{hcl.pals}.}
}
\value{
ggplot2 object with a scatter plot of selected gene.sets
}
\description{
This function allows to the user to examine the distribution of 
2 gene sets along the x.axis and y.axis. The color gradient
is generated using the a density estimate. See 
\href{https://github.com/LKremer/ggpointdensity}{ggpointdensity})
for more information.
}
\examples{
GS <- list(Bcells = c("MS4A1", "CD79B", "CD79A", "IGH1", "IGH2"),
           Tcells = c("CD3E", "CD3D", "CD3G", "CD7","CD8A"))
pbmc_small <- SeuratObject::pbmc_small
pbmc_small <- runEscape(pbmc_small, 
                        gene.sets = GS, 
                        min.size = NULL)
                        
scatterEnrichment(pbmc_small, 
                  assay = "escape",
                  x.axis = "Tcells",
                  y.axis = "Bcells")

}
